requirements.txt
fastapi
uvicorn
databases
sqlalchemy
pydantic[dotenv]
asyncpg
psycopg2-binary
starlette
passlib
bcrypt
python-multipart
PyJWT
gspread
httplib2
apiclient
google-api-python-client
oauth2client